completed EX_3.15 to 3.18 Phonebook database step  3 to 6

+++ part3/PhoneBook_Frontend/src/components/Persons.jsx
@@ -1,27 +1,33 @@
-
-
-
 import DeleteButton from "./DeleteButton";
 import axios from "axios";
+import { toast } from 'react-toastify'; 
 
 const Persons = ({ persons, setPersons}) => {
 
   const handleDelete = async (id) => {
-  try{
-   await axios.delete(`http://localhost:3000/api/persons/${id}`)
-   setPersons(persons.filter(person => person.id !== id))
-  }catch(error){
-    console.error("Error deleting Entries:", error);
-  }
-  }
+    if (!id) {
+      toast.error("Error: The ID is undefined.");
+      return;
+    }
+    console.log("Delete called with ID:", id);
+  
+    try {
+      await axios.delete(`http://localhost:3000/api/persons/${id}`);
+      setPersons(persons.filter(person => person._id !== id));
+      toast.success("Person deleted successfully!");
+    } catch (error) {
+      console.error("Error deleting Entries:", error);
+      toast.error("Failed to delete the person.");
+    }
+  };
   
   return (
     <div>
       {persons.map((person) => (
-        <div key={person.id}>
+        <div key={person._id}>
           <p>name: {person.name}</p>
           <p>mob: {person.number}</p>
-         <DeleteButton id={person.id} onDelete={() => handleDelete(person.id)}/>
+          <DeleteButton id={person._id} onDelete={handleDelete}/>
         </div>
       ))}
     </div>
@@ -29,4 +35,3 @@ const Persons = ({ persons, setPersons}) => {
 };
 
 export default Persons;
-
